Fix task 3! Improve table scheme in solution.sql

+++ clickhouse/solution.sql
@@ -2,9 +2,9 @@
 
 CREATE TABLE IF NOT EXISTS server_logs (
     timestamp DateTime,
-    user_id UInt16,
+    user_id UInt32,
     endpoint String,
-    response_time_ms UInt16,
+    response_time_ms UInt32,
     status_code UInt16
 ) ENGINE = MergeTree()
 ORDER BY (endpoint);
